Fixed memory corruption bug in python reader.
The python reader called sdf_close() once the reference count of
the SDF object reached zero. This would then deallocate all of the
memory allocated by the SDF library, including the arrays containing
the data.
For now, I have removed the sdf_close() call entirely so there is
now a massive memory leak. The obvious fix is to close the file once
the last reference to the file's contents has reached zero, but first
we need to return a data structure which tracks the returned items.