Added a floating-point consistency option.
If the code uses 80-bit precision arithmetic for the current
arrays, the answer will be inconsistent when run in parallel since
only 64-bits of precision is communicated to neighbouring processes.
This change adds a flag "CONS=1" that can be passed to to "make"
which enforces 64-bit arithmetic in the field solver.
Currently, it is only used for the Intel compiler.