Added "dump_first_after_restart" flag to io block
In the past, a "dump_first" flag in the output block would cause
an output dump immediately after restarting. Since this is rarely
the desired behaviour, the flag is now ignored when restarting.
To force a dump to occur immediately after restart, set
"dump_first_after_restart = T" in the output block.